refactor: add CustomCalendar component for date selection in student details
